Revert the refactoring of Card into Card data class and CardView display class. This simplifies the conceptualization and implementation greatly, and does so at only a minimal memory cost.
